+++ nautilus_trader/adapters/betfair/execution.py
@@ -275,6 +275,9 @@ class BetfairExecutionClient(LiveExecutionClient):
                     self._send_account_state(account_state)
                 except BetfairError as e:
                     self._log.warning(str(e))
+                    if "INVALID_SESSION_INFORMATION" in str(e):
+                        self._log.warning("Invalid session error, reconnecting...")
+                        await self._reconnect()
         except asyncio.CancelledError:
             self._log.debug("Canceled task 'update_account_state'")
         except Exception as e: