+++ RELEASES.md
@@ -71,6 +71,7 @@ Released on TBD (UTC).
 - Fixed gateway/TWS reconnect process for IBKR (#2710), thanks @bartlaw
 - Fixed Interactive Brokers options chain issue (#2711), thanks @FGU1
 - Fixed Partially filled bracket order and SL triggered for IBKR (#2704, #2717), thanks @bartlaw
+- Fixed instrument message decoding when no `exchange` value for Databento US equities
 - Restore task error logs for IBKR (#2716), thanks @bartlaw
 
 ### Documentation Updates