+++ RELEASES.md
@@ -72,8 +72,8 @@ Released on TBD (UTC).
 - Fixed Interactive Brokers options chain issue (#2711), thanks @FGU1
 - Fixed Partially filled bracket order and SL triggered for IBKR (#2704, #2717), thanks @bartlaw
 - Fixed instrument message decoding when no `exchange` value for Databento US equities
-- Restore task error logs for IBKR (#2716), thanks @bartlaw
 - Fixed fetching single-instrument trading fees for `Binance`, thanks @petioptrv
+- Restore task error logs for IBKR (#2716), thanks @bartlaw
 
 ### Documentation Updates
 None